+++ src/MicroElements.Swashbuckle.FluentValidation/FluentValidation/FluentValidationExtensions.cs
@@ -15,33 +15,6 @@ namespace MicroElements.Swashbuckle.FluentValidation
     public static class FluentValidationExtensions
     {
         /// <summary>
-        /// Contains <see cref="PropertyRule"/> and additional info.
-        /// </summary>
-        public readonly struct ValidationRuleContext
-        {
-            /// <summary>
-            /// PropertyRule.
-            /// </summary>
-            public readonly IValidationRule PropertyRule;
-
-            /// <summary>
-            /// Flag indication whether the <see cref="PropertyRule"/> is the CollectionRule.
-            /// </summary>
-            public readonly bool IsCollectionRule;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="ValidationRuleContext"/> struct.
-            /// </summary>
-            /// <param name="propertyRule">PropertyRule.</param>
-            /// <param name="isCollectionRule">Is a CollectionPropertyRule.</param>
-            public ValidationRuleContext(IValidationRule propertyRule, bool isCollectionRule)
-            {
-                PropertyRule = propertyRule;
-                IsCollectionRule = isCollectionRule;
-            }
-        }
-
-        /// <summary>
         /// Gets validation rules for validator.
         /// </summary>
         /// <param name="validator">Validator.</param>
@@ -67,19 +40,6 @@ namespace MicroElements.Swashbuckle.FluentValidation
         }
 
         /// <summary>
-        /// Returns property validators by property name ignoring name case.
-        /// </summary>
-        /// <param name="validator">Validator</param>
-        /// <param name="name">Property name.</param>
-        /// <returns>enumeration.</returns>
-        public static IEnumerable<IPropertyValidator> GetValidatorsForMemberIgnoreCase(this IValidator validator, string name)
-        {
-            return GetValidationRulesForMemberIgnoreCase(validator, name)
-                .SelectMany(propertyRule => propertyRule.PropertyRule.Components)
-                .OfType<IPropertyValidator>();
-        }
-
-        /// <summary>
         /// Returns all IValidationRules that are PropertyRule.
         /// If rule is CollectionPropertyRule then isCollectionRule set to true.
         /// </summary>
@@ -92,7 +52,8 @@ namespace MicroElements.Swashbuckle.FluentValidation
                 {
                     // CollectionPropertyRule<T, TElement> is also a PropertyRule.
                     var isCollectionRule = rule.GetType().Name.StartsWith("CollectionPropertyRule");
-                    return new ValidationRuleContext(rule, isCollectionRule);
+                    var reflectionContext = new ReflectionContext(propertyInfo: rule.Member);
+                    return new ValidationRuleContext(rule, isCollectionRule, reflectionContext);
                 });
         }